Polished Hub/Chirps

+++ app/Http/Controllers/ProductController.php
@@ -6,16 +6,28 @@ use App\Models\Product;
 use Illuminate\Http\Request;
 use Illuminate\Http\RedirectResponse;
 use App\Http\Requests\ProductRequest;
+use Illuminate\Support\Facades\Auth;
 
 class ProductController extends Controller
 {
     /**
      * Display a listing of the resource.
      */
-    public function index()
+    public function index(Request $request)
     {
+        $name = $request->input("name");
+        $stock = $request->input("stock");
+        $products = Product::when($name, fn ($query, $name) => $query->name($name));
+        $products = match ($stock) {
+            'available' => $products->onStock(),
+            'sold' => $products->offStock(),
+            default => $products,
+        };
+        $products = $products->withCount("comments")
+            ->latest()
+            ->paginate(12);
         return view('home', [
-            'products' => Product::withCount('comments')->latest()->paginate(12),
+            'products' => $products,
         ]);
     }
 
@@ -33,6 +45,7 @@ class ProductController extends Controller
     public function store(ProductRequest $request): RedirectResponse
     {
         $data = $request->validated();
+        $data['user_id'] = Auth::id();
         $product = Product::create($data);
         return redirect(route('products.show', ['product' => $product->id]))->with('success', 'Barang berhasil ditambahkan.');
     }
@@ -53,6 +66,9 @@ class ProductController extends Controller
      */
     public function edit(Product $product)
     {
+        if ($product->user_id !== Auth::id()) {
+            abort(403, 'Unauthorized action.');
+        }
         return view('edit', [
             'product' => $product,
         ]);
@@ -63,6 +79,9 @@ class ProductController extends Controller
      */
     public function update(ProductRequest $request, Product $product): RedirectResponse
     {
+        if ($product->user_id !== Auth::id()) {
+            abort(403, 'Unauthorized action.');
+        }
         $data = $request->validated();
         $product->update($data);
         return redirect(route('products.show', ['product' => $product->id]))->with('success', 'Barang berhasil diubah.');
@@ -73,7 +92,18 @@ class ProductController extends Controller
      */
     public function destroy(Product $product)
     {
+        if ($product->user_id !== Auth::id()) {
+            abort(403, 'Unauthorized action.');
+        }
         $product->delete();
         return redirect(route('products.index'))->with('success', 'Barang berhasil dihapus.');
+    }
+
+    public function myProducts()
+    {
+        $products = Product::where('user_id', Auth::id())->withCount("comments")
+            ->latest()
+            ->paginate(12);
+        return view('home', ['products' => $products]);
     }
 }
